+++ crates/ne_render/src/basic_cube.wgsl
@@ -1,7 +1,7 @@
 struct VertexOutput {
-    @location(0) tex_coord: vec2<f32>,
+    @location(0) tex_coords: vec2<f32>,
     @builtin(position) clip_position: vec4<f32>,
-};
+}
 struct Camera {
     //This is needed later ... 
     view_pos: vec4<f32>,
@@ -12,34 +12,43 @@ struct Camera {
 @binding(0)
 var<uniform> model_matrix: mat4x4<f32>;
 @group(0)
-@binding(2)
+@binding(3)
 var<uniform> camera: Camera;
 
 @vertex
 fn vs_main(
     @location(0) position: vec4<f32>,
-    @location(1) tex_coord: vec2<f32>,
+    @location(1) tex_coords: vec2<f32>,
 ) -> VertexOutput {
     let world_position = model_matrix * position;
 
     var result: VertexOutput;
     result.clip_position = camera.view_proj * world_position;
-    result.tex_coord = tex_coord;
+    result.tex_coords = tex_coords;
     return result;
 }
 
-@group(0)
-@binding(1)
-var r_color: texture_2d<u32>;
-
+@group(0) @binding(1)
+var t_diffuse: texture_2d<f32>;
+@group(0)@binding(2)
+var s_diffuse: sampler;
 @fragment
-fn fs_main(vertex: VertexOutput) -> @location(0) vec4<f32> {
-    let tex = textureLoad(r_color, vec2<i32>(vertex.tex_coord * 256.0), 0);
-    let v = f32(tex.x) / 255.0;
-    return vec4<f32>(1.0 - (v * 5.0), 1.0 - (v * 15.0), 1.0 - (v * 50.0), 1.0);
+fn fs_main(in: VertexOutput) -> @location(0) vec4<f32> {
+    return textureSample(t_diffuse, s_diffuse, in.tex_coords);
 }
 
-@fragment
-fn fs_wire(vertex: VertexOutput) -> @location(0) vec4<f32> {
-    return vec4<f32>(0.0, 0.5, 0.0, 0.5);
-}
+// @group(0)
+// @binding(1)
+// var r_color: texture_2d<u32>;
+
+// @fragment
+// fn fs_main(vertex: VertexOutput) -> @location(0) vec4<f32> {
+//     let tex = textureLoad(r_color, vec2<i32>(vertex.tex_coords * 256.0), 0);
+//     let v = f32(tex.x) / 255.0;
+//     return vec4<f32>(1.0 - (v * 5.0), 1.0 - (v * 15.0), 1.0 - (v * 50.0), 1.0);
+// }
+
+// @fragment
+// fn fs_wire(vertex: VertexOutput) -> @location(0) vec4<f32> {
+//     return vec4<f32>(0.0, 0.5, 0.0, 0.5);
+// }
